Split file up into its individual classes
